moved the register route function to a controller file
