+++ smartbrain-api/server.js
@@ -4,6 +4,8 @@ const bcrypt = require('bcrypt-nodejs');
 const cors = require('cors');
 const knex = require('knex');
 
+const register = require('./controllers/register');
+
 const db = knex({
   client: 'pg',
   connection: {
@@ -41,34 +43,7 @@ app.post('/signin', (req, res) => {
     .catch(err => res.status(400).json('Wrong credentials'))
 })
 
-app.post('/register', (req, res) => {
-  const { email, name, password } = req.body;
-  const hash = bcrypt.hashSync(password);
-
-  db.transaction(trx => {
-    trx.insert({
-      hash: hash,
-      email: email
-    })
-      .into('login')
-      .returning('email')
-      .then(loginEmail => {
-        return trx('users')
-          .returning('*')
-          .insert({
-            email: loginEmail[0],
-            name: name,
-            joined: new Date()
-          })
-          .then(user => {
-            res.json(user[0]);
-          })
-      })
-      .then(trx.commit)
-      .catch(trx.rollback)
-  })
-    .catch(err => res.status(400).json('unable to register'))
-})
+app.post('/register', (req, res) => { register.handleRegister(req, res, db, bcrypt) })
 
 app.get('/profile/:id', (req, res) => {
   const { id } = req.params;
